create addscore function
